requirements.txt
tensorflow==2.6.0
numpy
Pillow
